backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pydantic
torch
torch-directml
diffusers
transformers
accelerate
pillow
opencv-python
numpy
